refactor: out header to separate file
