+++ songbird/src/pages/quiz/quiz.js
@@ -1,0 +1,1 @@
+console.log('quiz');
